SentiStream/sentistream-updated/requirements.txt
apache-flink==1.17.0
gensim==4.3.1
nltk
numpy
pandas
redis
kafka-python
python-Levenshtein
matplotlib
scikit-learn
torch==1.13.1 --index-url https://download.pytorch.org/whl/cu118
